Dot rewritten to TS
