+++ src/components/Dot/Dot.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Dot from '@components/Dot'
+import { Dot } from '@components/Dot'
 
 # Dot
