+++ src/components/Dot/index.ts
@@ -1,0 +1,1 @@
+export * from './Dot'
