+++ src/components/index.ts
@@ -18,7 +18,7 @@ export { default as ColorNumber } from './ColorNumber'
 export * from './CustomLink'
 export { default as DateTimePicker } from './DateTimePicker'
 export { default as Divider } from './Divider'
-export { default as Dot } from './Dot'
+export * from './Dot'
 export { default as Dropdown } from './Dropdown'
 export { default as DropdownSelect } from './DropdownSelect'
 export { default as DropdownSelect2 } from './DropdownSelect2'
